# docs/openapi/specification.json
{
  "components": {
    "schemas": {
      "InvalidError": {
        "properties": {
          "error": {
            "type": "string"
          },
          "valids": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "type": "object"
      },
...
      "MetapoolChains": {
        "type": "string",
        "enum": ["avalanche", "bsc", "polygon", "arbitrum", "fantom"]
      },
...
      "MetapoolVolumeResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "float"
            }
          },
          "stats": {
            "type": "object",
            "properties": {
              "volume": {
                "type": "number",
                "format": "float"
              }
            }
          }
        }
      },
      "BridgeFilter": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "total": {
                  "type": "object",
                  "properties": {
                    "usd": {
                      "type": "number",
                      "format": "float"
                    }
                  }
                }
              },
              "additionalProperties": {
                "type": "object",
                "properties": {
                  "usd": {
                    "type": "number",
                    "format": "float"
                  },
                  "volume": {
                    "type": "number",
                    "format": "float"
                  }
                }
              }
            }
          },
          "stats": {
            "type": "object",
            "properties": {
              "usd": {
                "type": "object",
                "properties": {
                  "adjusted": {
                    "type": "number",
                    "format": "float"
                  },
                  "current": {
                    "type": "number",
                    "format": "float"
                  }
                }
              },
              "volume": {
                "type": "object",
                "additionalProperties": {
                  "type": "number",
                  "format": "float"
                }
              }
            }
          }
        },
        "example": {
          "data": {
            "2021-10-25": {
              "0xaa88c603d142c371ea0eac8756123c5805edee03": {
                "usd": 1355.478528122246,
                "volume": 123446.9082698944
              },
              "total": {
                "usd": 1355.478528122246
              }
            },
            "2021-10-26": {
              "0xaa88c603d142c371ea0eac8756123c5805edee03": {
                "usd": 1055.4443547315896,
                "volume": 100420
              },
              "total": {
                "usd": 1055.4443547315896
              }
            },
            "2021-10-27": {
              "0xaa88c603d142c371ea0eac8756123c5805edee03": {
                "usd": 757281.1329737306,
                "volume": 75000000
              },
              "total": {
                "usd": 757281.1329737306
              }
            },
            "2021-10-28": {
              "0xaa88c603d142c371ea0eac8756123c5805edee03": {
                "usd": 265722.62781034486,
                "volume": 15805181.048415447
              },
              "total": {
                "usd": 265722.62781034486
              }
            },
            "2021-10-29": {
              "0xaa88c603d142c371ea0eac8756123c5805edee03": {
                "usd": 626228.3000757314,
                "volume": 19759417.301290344
              },
              "total": {
                "usd": 626228.3000757314
              }
            },
            "2021-10-30": {
              "0xaa88c603d142c371ea0eac8756123c5805edee03": {
                "usd": 368915.2027302926,
                "volume": 12501172.662217114
              },
              "total": {
                "usd": 368915.2027302926
              }
            }
          },
          "stats": {
            "usd": {
              "adjusted": 2020558.186472953,
              "current": 3624643.8468636745
            },
            "volume": {
              "0xaa88c603d142c371ea0eac8756123c5805edee03": 123289637.9201928
            }
          }
        }
      }
    }
  },
  "info": {
    "title": "SYN Tracker",
    "version": "1.0.0",
    "description": "An API to track useful things about SYN. Still a WIP. View the repo [here](https://github.com/blazewashere/syn-tracker-api)",
    "contact": {
      "email": "blaze@dorime.org"
    },
    "license": {
      "name": "Boost Software License 1.0",
      "url": " https://www.boost.org/LICENSE_1_0.txt"
    }
  },
  "openapi": "3.0.3",
  "servers": [
    { "url": "https://synapse.dorime.org", "description": "production" },
    { "url": "http://localhost:1337", "description": "development" }
  ],
  "paths": {
...
    "/api/v1/analytics/volume/metapool/{chain}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "chain",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MetapoolChains"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetapoolVolumeResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "volume of nUSD on a chain metapool",
        "tags": ["Metapool Volume"]
      }
    },
...
    "/api/v1/analytics/volume/bsc/filter/{token}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["nusd", "syn", "high", "dog"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BridgeFilter" }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "volume of a token on the BSC bridge",
        "tags": ["Bridge Volume"]
      }
    },
...
    "/api/v1/analytics/volume/ethereum/filter/{token}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["nusd", "syn", "high", "dog", "usdt", "usdc", "dai"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BridgeFilter" }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "volume of a token on the ETH bridge",
        "tags": ["Bridge Volume"]
      }
    },
...
    "/api/v1/analytics/volume/polygon/filter/{token}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["nusd", "syn"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BridgeFilter" }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "volume of a token on the MATIC/Polygon bridge",
        "tags": ["Bridge Volume"]
      }
    },
...
  }
}